Schedule model / double -> bigdecimal

+++ src/main/java/com/warr/ferr/model/Schedule.java
@@ -19,14 +19,14 @@ public class Schedule {
     private String eventTitle;
     private Timestamp eventStartDate;
     private Timestamp eventEndDate;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private Timestamp createdAt;
     private Date promiseDate;
     private String startLocation;
 
     // 생성자를 포함하여, 필요한 메서드 추가 가능
-    public Schedule(int userId, String contentId, String eventTitle, Timestamp eventStartDate, Timestamp eventEndDate, Double latitude, Double longitude, Timestamp createdAt, Date promiseDate, String startLocation) {
+    public Schedule(int userId, String contentId, String eventTitle, Timestamp eventStartDate, Timestamp eventEndDate, BigDecimal latitude, BigDecimal longitude, Timestamp createdAt, Date promiseDate, String startLocation) {
         this.userId = userId;
         this.contentId = contentId;
         this.eventTitle = eventTitle;
